videos/requirements.txt
grpcio
grpcio-tools
psycopg2-binary